小更新

+++ main.py
@@ -21,9 +21,9 @@ if __name__ == "__main__":
             utils.asyncio.run(main())
         except KeyboardInterrupt:
             logger.info("聊天记录已被删除")
-            utils.os.remove("./cache/chat.json")
+            if utils.os.path.exists("./cache/chat.json"):
+                utils.os.remove("./cache/chat.json")
             break
         except Exception as e:
             logger.error(e)
-            logger.info("你可以再次启动程序，聊天记录已经保存了")
-            break
+            logger.info("程序异常，已重启")
